Merge pull request #2 from adamkoziol/master

+++ fileExtractionProcessing.py
@@ -24,6 +24,7 @@ def make_path(inPath):
         if exception.errno != errno.EEXIST:
             raise
 
+
 def unzipping():
     zipFile = glob.glob("*.zip")
     if not zipFile:
@@ -46,17 +47,20 @@ def moveExtract(strain, gzFiles, path):
     """Renames, moves, and uncompresses .gz files"""
     forward = str(strain) + "_R1_001.fastq.gz"
     reverse = str(strain) + "_R2_001.fastq.gz"
-    shutil.move(gzFiles[0], "%s/%s/%s" % (path, strain, forward))
-    shutil.move(gzFiles[1], "%s/%s/%s" % (path, strain, reverse))
-    sys.stdout.write('.')
-    gzipCommandForward = "gzip -d %s/%s/%s" % (path, strain, forward)
-    gzipCommandReverse = "gzip -d %s/%s/%s" % (path, strain, reverse)
-    subprocess.call(gzipCommandForward, shell=True, stdout=open(os.devnull, 'wb'), stderr=open(os.devnull, 'wb'))
-    subprocess.call(gzipCommandReverse, shell=True, stdout=open(os.devnull, 'wb'), stderr=open(os.devnull, 'wb'))
+    print strain
+    if os.path.isfile(gzFiles[0]) and os.path.isfile(gzFiles[1]):
+        shutil.move(gzFiles[0], "%s/%s/%s" % (path, strain, forward))
+        shutil.move(gzFiles[1], "%s/%s/%s" % (path, strain, reverse))
+        sys.stdout.write('.')
+        gzipCommandForward = "gzip -d %s/%s/%s" % (path, strain, forward)
+        gzipCommandReverse = "gzip -d %s/%s/%s" % (path, strain, reverse)
+        subprocess.call(gzipCommandForward, shell=True, stdout=open(os.devnull, 'wb'), stderr=open(os.devnull, 'wb'))
+        subprocess.call(gzipCommandReverse, shell=True, stdout=open(os.devnull, 'wb'), stderr=open(os.devnull, 'wb'))
 
 
 def folderer(sampleNames, path):
     """Move the .gz files into appropriately named folders, and decompress them"""
+    os.chdir(path)
     gzCheck = glob.glob("*.gz")
     if not gzCheck:
         print("Processing files.")
@@ -66,10 +70,11 @@ def folderer(sampleNames, path):
             if not gzFiles:
                 pass
             else:
-                moveExtract(strain, gzFiles, path)
+                moveExtract(strain, sorted(gzFiles), path)
     else:
         print("Moving and extracting fastq files.")
         for strain in sampleNames:
+            os.chdir(path)
             # Make the required folders (if necessary)
             make_path("%s/%s" % (path, strain))
             # Get the .gz files into a list
@@ -80,9 +85,9 @@ def folderer(sampleNames, path):
                 if not gzFiles:
                     pass
                 else:
-                    moveExtract(strain, gzFiles, path)
+                    moveExtract(strain, sorted(gzFiles), path)
             else:
-                moveExtract(strain, gzFiles, path)
+                moveExtract(strain, sorted(gzFiles), path)
 
 
 def functionsGoNOW(sampleNames, path):
